refactor(clp-s): Split some of `Output`'s responsibilities into a new `QueryRunner` class. (#817)
