+++ components/core/CMakeLists.txt
@@ -327,6 +327,8 @@ set(SOURCE_FILES_clp_s_unitTest
     src/clp_s/search/OutputHandler.hpp
     src/clp_s/search/Projection.cpp
     src/clp_s/search/Projection.hpp
+    src/clp_s/search/QueryRunner.cpp
+    src/clp_s/search/QueryRunner.hpp
     src/clp_s/search/SchemaMatch.cpp
     src/clp_s/search/SchemaMatch.hpp
     src/clp_s/TimestampDictionaryReader.cpp
