+++ components/core/src/clp_s/SchemaReader.cpp
@@ -258,11 +258,11 @@ bool SchemaReader::get_next_message_with_metadata(
 }
 
 void SchemaReader::initialize_filter(FilterClass* filter) {
-    filter->init(this, m_schema_id, m_columns);
+    filter->init(this, m_columns);
 }
 
 void SchemaReader::initialize_filter_with_column_map(FilterClass* filter) {
-    filter->init(this, m_schema_id, m_column_map);
+    filter->init(this, m_column_map);
 }
 
 void SchemaReader::generate_local_tree(int32_t global_id) {
