+++ components/core/src/clp_s/clp-s.cpp
@@ -175,8 +175,11 @@ bool search_archive(
     }
 
     // Narrow against schemas
-    SchemaMatch match_pass(archive_reader->get_schema_tree(), archive_reader->get_schema_map());
-    if (expr = match_pass.run(expr); std::dynamic_pointer_cast<ast::EmptyExpr>(expr)) {
+    auto match_pass = std::make_shared<SchemaMatch>(
+            archive_reader->get_schema_tree(),
+            archive_reader->get_schema_map()
+    );
+    if (expr = match_pass->run(expr); std::dynamic_pointer_cast<ast::EmptyExpr>(expr)) {
         SPDLOG_INFO("No matching schemas for query '{}'", query);
         return true;
     }
@@ -262,7 +265,6 @@ bool search_archive(
             match_pass,
             expr,
             archive_reader,
-            timestamp_dict,
             std::move(output_handler),
             command_line_arguments.get_ignore_case()
     );
